+++ SIMSclient/src/presentation/generalmanagerui/controller/PromotionSumController.java
@@ -3,6 +3,7 @@ package presentation.generalmanagerui.controller;
 import java.rmi.RemoteException;
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.Optional;
 
 import bussinesslogic.commoditybl.CommodityController;
 import bussinesslogic.promotionbl.PromotionSumBL;
@@ -18,6 +19,7 @@ import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.Label;
@@ -26,6 +28,7 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableColumn.CellEditEvent;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.util.Callback;
 import presentation.common.EditingCellDate;
@@ -97,6 +100,10 @@ public class PromotionSumController extends PromotionMakingController{
 
 	@FXML
 	public void find() throws RemoteException{
+		if(findingField.getText()==null||findChoice.getValue()==null){
+			Alert warning = new Alert(Alert.AlertType.WARNING,"请填写好查询信息");
+			warning.showAndWait();
+		}else{
 		ArrayList<PromotionTotalVO> list = service.find(findingField.getText(),FindPromotionType.getType(findChoice.getValue()));
 	       if(list==null){
 	    	   Alert error = new Alert(Alert.AlertType.WARNING,ResultMessage.NOTFOUND.value);
@@ -107,12 +114,19 @@ public class PromotionSumController extends PromotionMakingController{
 	    	   table.getItems().addAll(list);
 	    	   initFind();
 	       }
+		}
 	}
 
 
 
 	@FXML
 	public void insert() throws RemoteException{
+		if(startPicker.getValue()==null||endPicker.getValue()==null||sumField.getText()==null||
+				voucherField.getText()==null){
+			Alert warning = new Alert(Alert.AlertType.WARNING,"请填写好所有的信息");
+			warning.showAndWait();
+		}
+		else{
 		ArrayList<GiftVO> gifts = null;
 		 PromotionTotalVO vo = new PromotionTotalVO(idLabel.getText(),startPicker.getValue(),endPicker.getValue(),
 				 Double.parseDouble(sumField.getText()),Double.parseDouble(voucherField.getText()),gifts);
@@ -133,14 +147,25 @@ public class PromotionSumController extends PromotionMakingController{
 					}
 	    	    }
 	    	});
+		}
 	}
 
 	@FXML
 	public void insertGift() throws RemoteException{
+		if(giftChoice.getValue()==null||numberField.getText()==null){
+			Alert warning = new Alert(Alert.AlertType.WARNING,"请填写好所有的信息");
+			warning.showAndWait();
+		}
+		else if(currentPromotion==null){
+			Alert warning = new Alert(Alert.AlertType.WARNING,"请点击查看赠品列表选择需要赠品的策略");
+			warning.showAndWait();
+		}
+		else{
 		 GiftVO vo = new GiftVO(giftChoice.getValue(),Integer.parseInt(numberField.getText()));
 	     giftList.add(vo);
 	     giftTable.setItems(giftList);
 	     updateGiftList();
+		}
 	}
 
 
@@ -321,13 +346,18 @@ public class PromotionSumController extends PromotionMakingController{
                         this.setGraphic(delBtn);
                         delBtn.setOnMouseClicked((me) -> {
                         	PromotionTotalVO clickedUser = this.getTableView().getItems().get(this.getIndex());
-                            try {
-								service.delete(clickedUser);
-							} catch (RemoteException e) {
-								e.printStackTrace();
-							}
-                            list.remove(clickedUser);
-                            table.setItems(list);
+                        	 try {
+	                            	Alert alert = new Alert(AlertType.CONFIRMATION);
+	                            	alert.setContentText("确认删除？");
+	                            	Optional<ButtonType> result = alert.showAndWait();
+	                            	if (result.get() == ButtonType.OK){
+	                            		service.delete(clickedUser);
+	                            		  list.remove(clickedUser);
+	      	                              table.setItems(list);
+	                            	}
+								} catch (RemoteException e) {
+									e.printStackTrace();
+								}
                         });
                     }
                 }
